requirements.txt
# Core Flask framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
WTForms==3.0.1
Jinja2==3.1.2

# WSGI server for production
gunicorn==21.2.0

# Database drivers
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# AI APIs
anthropic==0.57.1
openai==1.98.0

# Document processing
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.0.1
openpyxl==3.1.2
python-magic==0.4.27

# Security and authentication
Werkzeug==2.3.7
bcrypt==4.0.1
cryptography==41.0.7
PyJWT==2.8.0

# HTTP requests and utilities
requests==2.31.0
urllib3==2.0.7
python-dateutil==2.8.2
python-dotenv==1.0.0

# File handling
filetype==1.2.0

# Azure integration
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0

# Caching and background tasks
redis==5.0.1
celery==5.3.4
flask-caching==2.1.0

# Celery monitoring (optional)
flower==2.0.1

# Monitoring and health checks
flask-healthz==0.0.3
flask-limiter==3.5.0

# Input validation
marshmallow==3.20.1

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
